all good

+++ Assets/Scripts/Level/Editor/LevelManagerEditor.cs
@@ -1,20 +1,20 @@
-using UnityEngine;
-using UnityEditor;
 using LevelManaging;
+using UnityEditor;
+using UnityEngine;
 
 [CustomEditor(typeof(LevelManager))]
 public class LevelManagerEditor : Editor
 {
-    private void OnEnable() 
-    {
-        (target as LevelManager).LoadLevelNumberFromPrefs();
-    }
-    public override void OnInspectorGUI() {
-        base.OnInspectorGUI();
+	private void OnEnable()
+	{
+		(target as LevelManager).LoadLevelNumberFromPrefs();
+	}
 
-        EditorGUILayout.Space(10);
-        if (GUILayout.Button("Clear playerPrefs", GUILayout.Height(25))) {
-            PlayerPrefs.DeleteAll();
-        }
-    }
+	public override void OnInspectorGUI()
+	{
+		base.OnInspectorGUI();
+
+		EditorGUILayout.Space(10);
+		if (GUILayout.Button("Clear playerPrefs", GUILayout.Height(25))) PlayerPrefs.DeleteAll();
+	}
 }
